Switch to Recoil for state management (removed support for multiple solutions per puzzle)
Now each puzzle is hardcoded to have a single solution. This might be an annoying limitation at some point, but for now we should keep it simple. This removes the need for a ton of logic around creating, renaming, copying, etc. solutions. It also simplifies the data store. If we see the need for multiple solutions we can also go to the fixed number of solutions methodology that TIS-100 uses.